Auto stash before merge of "master" and "origin/master"

+++ BDD/BDD_PI.sql
@@ -65,9 +65,9 @@ CREATE TABLE Article(
   Label INT(11) NOT NULL,
   Fermier VARCHAR(45),
   Statut ENUM('A vendre', 'Achete', 'Vendu','Partiellement achete'),
-  Qte_article INT(4),
-  Qte_restantes INT(4),
-  Qte_acheter INT(4),
+  Qte_article float,
+  Qte_restantes float,
+  Qte_acheter float,
   PRIMARY KEY (id_article),
   FOREIGN KEY (Label) REFERENCES Label (id_label) on delete cascade
 );
